Fix error with bandit >= 1.4.0
Bandit 1.4.0 removed `status`, `method` and `request_target` from
telemetry metadata since they can be obtained from the `conn` struct.
This was crashing the lib. The changelog of the removal can be seen
[here](https://github.com/mtrudel/bandit/blob/main/CHANGELOG.md#140-26-mar-2024).
This commit isn't a breaking change since we are just grabbing the
fields straight from `conn` which was always present.

I've also updated the deps, added credo and Github Action to run format,
lint, tests and dialyzer on PRs and merges.